moving endpoint loop out of module
