+++ WordpressSync/processEndpoints.js
@@ -1,9 +1,4 @@
 const GitHub = require('github-api');
-const endpoints = require('./endpoints.json').data;
-const committer = {
-  name: process.env["GITHUB_NAME"],
-  email: process.env["GITHUB_EMAIL"]
-};
 const commitTitlePosts = 'Wordpress Posts Update';
 const commitTitlePages = 'Wordpress Pages Update';
 const commitTitleMedia = 'Wordpress Media Update';
@@ -191,163 +186,164 @@ const syncBinaryFile = async (source_url, gitRepo, mediaTree, endpoint) => {
   treeNode.sha = sha;
 };
 
-const doProcessEndpoints = async () => {
-  const debugMode = process.env.debug?.toLowerCase()==='true';
-  const gitModule = new GitHub({ token: process.env["GITHUB_TOKEN"] });
+/**
+ * process a Wordpress endpoint and place the data in GitHub
+ * @param {{WordPressUrl: string, GitHubTarget: {Owner: string, Repo: string, Path: string,Branch: string}}} endpoint 
+ * @param {{token:string}} gitHubCredentials 
+ * @param {{name:string,email:string}} gitHubCommitter 
+ */
+const SyncEndpoint = async (endpoint, gitHubCredentials, gitHubCommitter) => {
+  const gitModule = new GitHub(gitHubCredentials);
+  const wordPressApiUrl = endpoint.WordPressUrl+apiPath;
+  const gitRepo = await gitModule.getRepo(endpoint.GitHubTarget.Owner,endpoint.GitHubTarget.Repo);
 
-  for(const endpoint of endpoints.projects.filter(x=>debugMode && x.enabledLocal || !debugMode && x.enabled)) {
-    console.log(`*** Checking endpoint for ${endpoint.name} ***`);
-    const wordPressApiUrl = endpoint.WordPressUrl+apiPath;
-    const gitRepo = await gitModule.getRepo(endpoint.GitHubTarget.Owner,endpoint.GitHubTarget.Repo);
+  //List of WP categories
+  const categorylist = await fetchDictionary(wordPressApiUrl,'categories');
+  const taglist = await fetchDictionary(wordPressApiUrl,'tags');
+  const userlist = await fetchDictionary(wordPressApiUrl,'users');
 
-    //List of WP categories
-    const categorylist = await fetchDictionary(wordPressApiUrl,'categories');
-    const taglist = await fetchDictionary(wordPressApiUrl,'tags');
-    const userlist = await fetchDictionary(wordPressApiUrl,'users');
-
-    const postMap = new Map();
-    const pagesMap = new Map();
-    const mediaMap = endpoint.GitHubTarget.SyncMedia ? new Map() : null;
-    
-    // MEDIA
-    const mediaContentPlaceholder = 'TBD : Binary file to be updated in a later step';
-    if(endpoint.GitHubTarget.SyncMedia) {
-      const allMedia = await WpApi_GetPagedData(wordPressApiUrl,'media');
-
-      allMedia.forEach(x=>{
-        const jsonData = getWpCommonJsonData(x,userlist);
-        delete jsonData.excerpt;
+  const postMap = new Map();
+  const pagesMap = new Map();
+  const mediaMap = endpoint.GitHubTarget.SyncMedia ? new Map() : null;
   
-        if(x.media_details.sizes) {
-          jsonData.sizes = Object.keys(x.media_details.sizes).map(s=>({
-            type:s,
-            path:pathFromMediaSourceUrl(x.media_details.sizes[s].source_url),
-            ...x.media_details.sizes[s]}));
+  // MEDIA
+  const mediaContentPlaceholder = 'TBD : Binary file to be updated in a later step';
+  if(endpoint.GitHubTarget.SyncMedia) {
+    const allMedia = await WpApi_GetPagedData(wordPressApiUrl,'media');
 
-          jsonData.sizes.sort((a,b)=>b.width-a.width); //Big first
+    allMedia.forEach(x=>{
+      const jsonData = getWpCommonJsonData(x,userlist);
+      delete jsonData.excerpt;
 
-          //put binary placeholders so they aren't deleted.  Will search for these if an update happens.
-          for (const s of jsonData.sizes) {
-            mediaMap.set(s.path, mediaContentPlaceholder);
+      if(x.media_details.sizes) {
+        jsonData.sizes = Object.keys(x.media_details.sizes).map(s=>({
+          type:s,
+          path:pathFromMediaSourceUrl(x.media_details.sizes[s].source_url),
+          ...x.media_details.sizes[s]}));
+
+        jsonData.sizes.sort((a,b)=>b.width-a.width); //Big first
+
+        //put binary placeholders so they aren't deleted.  Will search for these if an update happens.
+        for (const s of jsonData.sizes) {
+          mediaMap.set(s.path, mediaContentPlaceholder);
+        }
+      } else {
+        //PDF
+        jsonData.path = pathFromMediaSourceUrl(x.source_url);
+        mediaMap.set(jsonData.path, mediaContentPlaceholder);
+      }
+
+      mediaMap.set(`${pathFromMediaSourceUrl(x.source_url).split('.')[0]}.json`,wrapInFileMeta(endpoint,jsonData));
+    });
+
+    let mediaTree = await createTreeFromFileMap(gitRepo,endpoint.GitHubTarget.Branch,mediaMap,endpoint.GitHubTarget.MediaPath);
+ 
+    const mediaChanges = mediaTree
+      .filter(x=>x.content && x.content!==mediaContentPlaceholder)
+      .map(mt=>JSON.parse(mt.content).data);
+
+    if(mediaChanges.length) {
+      console.log(`Checking ${mediaTree.length} media items`);
+
+      //Pull in binaries for any media meta changes
+      for (const mediaTreeItem of mediaChanges) {
+        if (mediaTreeItem.sizes) {
+          //Sized images
+          for (const sizeJson of mediaTreeItem.sizes) {
+            //sometimes the source_url is full and sometimes it is relative
+            const imageUrl = (sizeJson.source_url.startsWith('http') ? '' : endpoint.WordPressUrl) + sizeJson.source_url;
+            await syncBinaryFile(imageUrl, gitRepo, mediaTree, endpoint);
           }
         } else {
-          //PDF
-          jsonData.path = pathFromMediaSourceUrl(x.source_url);
-          mediaMap.set(jsonData.path, mediaContentPlaceholder);
+          //not sized media (PDF or non-image)
+          await syncBinaryFile(mediaTreeItem.wordpress_url,gitRepo, mediaTree, endpoint);
         }
+      }
+    }
 
-        mediaMap.set(`${pathFromMediaSourceUrl(x.source_url).split('.')[0]}.json`,wrapInFileMeta(endpoint,jsonData));
+    //Remove any leftover binary placeholders...
+    mediaTree = mediaTree.filter(x=>x.content !== mediaContentPlaceholder);
+    
+    await PrIfChanged(gitRepo, endpoint.GitHubTarget.Branch, mediaTree, `${commitTitleMedia} (${mediaTree.length} updates)`, gitHubCommitter, true);
+  }
+  
+  /**
+   * Places the media section if SyncMedia is on
+   * @param {{}} jsonData 
+   * @param {{}} WpRow 
+   * @param {string} HTML 
+   */
+  const addMediaSection = (jsonData,WpRow,HTML) => {
+    if(endpoint.GitHubTarget.SyncMedia) {
+      if(WpRow.featured_media) {
+        jsonData.featured_media = WpRow.featured_media;
+      }
+
+      jsonData.media = [];
+      mediaMap.forEach(m=>{
+        //Look at media JSON only
+        if(m.data && m.data.sizes) {
+          m.data.sizes.forEach(s=>{
+            const source_url_match = HTML.includes(s.source_url);
+            const featured = jsonData.featured_media===m.data.id;
+            
+            if(featured || source_url_match) {
+              jsonData.media.push({
+                id:m.data.id,
+                ...s,
+                source_url_match,
+                featured
+              });
+            }
+          });
+        }
       });
 
-      let mediaTree = await createTreeFromFileMap(gitRepo,endpoint.GitHubTarget.Branch,mediaMap,endpoint.GitHubTarget.MediaPath);
-   
-      const mediaChanges = mediaTree
-        .filter(x=>x.content && x.content!==mediaContentPlaceholder)
-        .map(mt=>JSON.parse(mt.content).data);
-
-      if(mediaChanges.length) {
-        console.log(`Checking ${mediaTree.length} media items`);
-
-        //Pull in binaries for any media meta changes
-        for (const mediaTreeItem of mediaChanges) {
-          if (mediaTreeItem.sizes) {
-            //Sized images
-            for (const sizeJson of mediaTreeItem.sizes) {
-              //sometimes the source_url is full and sometimes it is relative
-              const imageUrl = (sizeJson.source_url.startsWith('http') ? '' : endpoint.WordPressUrl) + sizeJson.source_url;
-              await syncBinaryFile(imageUrl, gitRepo, mediaTree, endpoint);
-            }
-          } else {
-            //not sized media (PDF or non-image)
-            await syncBinaryFile(mediaTreeItem.wordpress_url,gitRepo, mediaTree, endpoint);
-          }
-        }
+      //Remove empty media array
+      if (!jsonData.media.length) {
+        delete jsonData.media;
       }
-
-      //Remove any leftover binary placeholders...
-      mediaTree = mediaTree.filter(x=>x.content !== mediaContentPlaceholder);
-      
-      await PrIfChanged(gitRepo, endpoint.GitHubTarget.Branch, mediaTree, `${commitTitleMedia} (${mediaTree.length} updates)`, committer, true);
     }
-    
-    /**
-     * Places the media section if SyncMedia is on
-     * @param {{}} jsonData 
-     * @param {{}} WpRow 
-     * @param {string} HTML 
-     */
-    const addMediaSection = (jsonData,WpRow,HTML) => {
-      if(endpoint.GitHubTarget.SyncMedia) {
-        if(WpRow.featured_media) {
-          jsonData.featured_media = WpRow.featured_media;
-        }
+  };
+  
+  // POSTS
+  const allPosts = await WpApi_GetPagedData(wordPressApiUrl,'posts');
+  allPosts.forEach(x=>{
+    const jsonData = getWpCommonJsonData(x,userlist);
+    jsonData.categories = x.categories.map(t=>categorylist[t]);
+    jsonData.tags = x.tags.map(t=>taglist[t]);
 
-        jsonData.media = [];
-        mediaMap.forEach(m=>{
-          //Look at media JSON only
-          if(m.data && m.data.sizes) {
-            m.data.sizes.forEach(s=>{
-              const source_url_match = HTML.includes(s.source_url);
-              const featured = jsonData.featured_media===m.data.id;
-              
-              if(featured || source_url_match) {
-                jsonData.media.push({
-                  id:m.data.id,
-                  ...s,
-                  source_url_match,
-                  featured
-                });
-              }
-            });
-          }
-        });
+    const HTML = cleanupContent(x.content.rendered);
+  
+    addMediaSection(jsonData,x,HTML);
 
-        //Remove empty media array
-        if (!jsonData.media.length) {
-          delete jsonData.media;
-        }
-      }
-    };
-    
-    // POSTS
-    const allPosts = await WpApi_GetPagedData(wordPressApiUrl,'posts');
-    allPosts.forEach(x=>{
-      const jsonData = getWpCommonJsonData(x,userlist);
-      jsonData.categories = x.categories.map(t=>categorylist[t]);
-      jsonData.tags = x.tags.map(t=>taglist[t]);
+    postMap.set(`${x.slug}.json`,wrapInFileMeta(endpoint,jsonData));
+    postMap.set(`${x.slug}.html`,HTML);
+  });
 
-      const HTML = cleanupContent(x.content.rendered);
-    
-      addMediaSection(jsonData,x,HTML);
-
-      postMap.set(`${x.slug}.json`,wrapInFileMeta(endpoint,jsonData));
-      postMap.set(`${x.slug}.html`,HTML);
-    });
-
-    const postTree = await createTreeFromFileMap(gitRepo,endpoint.GitHubTarget.Branch,postMap,endpoint.GitHubTarget.PostPath);
-    await PrIfChanged(gitRepo, endpoint.GitHubTarget.Branch, postTree, `${commitTitlePosts} (${postTree.filter(x=>x.path.endsWith(".html")).length} updates)`, committer, true);
+  const postTree = await createTreeFromFileMap(gitRepo,endpoint.GitHubTarget.Branch,postMap,endpoint.GitHubTarget.PostPath);
+  await PrIfChanged(gitRepo, endpoint.GitHubTarget.Branch, postTree, `${commitTitlePosts} (${postTree.filter(x=>x.path.endsWith(".html")).length} updates)`, gitHubCommitter, true);
 
 
-    // PAGES
-    const allPages = await WpApi_GetPagedData(wordPressApiUrl,'pages');
-    allPages.forEach(x=>{
-      const jsonData = getWpCommonJsonData(x,userlist);
-      jsonData.parent = x.parent;
-      jsonData.menu_order = x.menu_order;
+  // PAGES
+  const allPages = await WpApi_GetPagedData(wordPressApiUrl,'pages');
+  allPages.forEach(x=>{
+    const jsonData = getWpCommonJsonData(x,userlist);
+    jsonData.parent = x.parent;
+    jsonData.menu_order = x.menu_order;
 
-      const HTML = cleanupContent(x.content.rendered);
+    const HTML = cleanupContent(x.content.rendered);
 
-      addMediaSection(jsonData,x,HTML);
+    addMediaSection(jsonData,x,HTML);
 
-      pagesMap.set(`${x.slug}.json`,wrapInFileMeta(endpoint,jsonData));
-      pagesMap.set(`${x.slug}.html`,HTML);
-    });
+    pagesMap.set(`${x.slug}.json`,wrapInFileMeta(endpoint,jsonData));
+    pagesMap.set(`${x.slug}.html`,HTML);
+  });
 
-    const pagesTree = await createTreeFromFileMap(gitRepo,endpoint.GitHubTarget.Branch,pagesMap,endpoint.GitHubTarget.PagePath);
-    await PrIfChanged(gitRepo, endpoint.GitHubTarget.Branch, pagesTree, `${commitTitlePages} (${pagesTree.filter(x=>x.path.endsWith(".html")).length} updates)`, committer, true);
-  }
+  const pagesTree = await createTreeFromFileMap(gitRepo,endpoint.GitHubTarget.Branch,pagesMap,endpoint.GitHubTarget.PagePath);
+  await PrIfChanged(gitRepo, endpoint.GitHubTarget.Branch, pagesTree, `${commitTitlePages} (${pagesTree.filter(x=>x.path.endsWith(".html")).length} updates)`, gitHubCommitter, true);
 };
 
 module.exports = {
-  doProcessEndpoints
+  SyncEndpoint
 };
